Add dsvm-functional tox env to fix functional job
Fixes for the bug in question landed in master to get the functional
job working again, and the first step in backporting those fixes is
ensuring that the {gate|check}-neutron-dsvm-functional jobs can invoke
the dsvm-functional tox env.  There is no good backport for this
addition because the patches in master are highly coupled so as to
make cherry-picking basically impossible without this patch.

Change-Id: Ide15f6925d516c18cae1afa98f5661f0609bb50e
Partial-bug: #1336172