Break out config and entry point out of l3/agent file
This is done in order to show the split between a main() and
the agent's inner workings. At the same time the L3 agent for
Neutron is also being restructured and to this aim, its config
is also moved out of the agent file.

Keep main() for the time being until we get dependent
services updated.

Partially-Implements: bp restructure-l3-agent
Partially-Implements: bp core-vendor-decomposition

Change-Id: I0d07d91ba301ee1aa51dabcf964a96edc0d6a3e0